+++ lib/pages/HomePage.dart
@@ -10,7 +10,10 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  final _controller = TextEditingController();
+  // variables for saving information so can send it later
+  GlobalKey<FormState> _key = new GlobalKey();
+  bool _validate = false;
+  String firstName, lastName, cardNumber, city, state, phoneNumber;
 
   ///designs the style for the App Bar on the homepage
   Widget _appBar() {
@@ -48,81 +51,368 @@ class _HomePageState extends State<HomePage> {
       centerTitle: true,
     );
   }
-  
-  ///styling for the TextInputForm
-  Widget _textInputForm(String error, String label, IconData icon) {
-    return Container(
-      margin: ThemeStyle.marginALL,
-      // actual TextForm Field
-      child: TextFormField(
-        controller: _controller,
-        //validator creates a condition to be checked for when inputting response
-        validator: (value) {
-          return value.isEmpty ? error : null;
-        },
-        //creates the decoration for the input box
-        decoration: InputDecoration(
-          //outside border details
-          enabledBorder: OutlineInputBorder(
-            borderSide: BorderSide(
-              color: ThemeStyle.greyColor,
-              width: 2,
-            ),
-            borderRadius: BorderRadius.all(Radius.circular(30)),
-          ),
-          //label text details
-          alignLabelWithHint: true,
-          labelText: label,
-          labelStyle: ThemeStyle.formLabelText,
-          //icon details
-          prefixIcon: Icon(
-            icon,
-            size: ThemeStyle.iconSize,
-            color: ThemeStyle.blackColor,
-          ),
-          //details when selecting that input box
-          focusedBorder: OutlineInputBorder(
-            borderSide: BorderSide(
-              color: ThemeStyle.primColor,
-              width: 2,
-            ),
-            borderRadius: ThemeStyle.borderRadiusStyle,
-          ),
-          //filling in the color of the inside input box
-          fillColor: ThemeStyle.lightGreyColor,
-          filled: true,
-        ),
-      ),
-    );
-  }
 
-  @override
-  Widget build(BuildContext context) {
-    //helps control the screen sizing 
-    ScreenUtil.init();
-    ScreenUtil.init(width: 750, height: 1334, allowFontScaling: true);
-
-    return Scaffold(
-      appBar: _appBar(),
-      resizeToAvoidBottomPadding: true,
-      body: Center(
-        child: ListView(
-          padding: ThemeStyle.marginALL,
-          children: <Widget>[
-            Center(
+   ///login form -- includes header, input options, and button
+  Widget _formUI() {
+    return Padding(
+      padding: const EdgeInsets.only(top: 5.0, bottom: 5, left: 20, right: 20),
+      child: new Column(
+        children: <Widget>[
+          // header
+          Padding(
+            padding: const EdgeInsets.only(bottom: 16),
+            child: Center(
               child: Text(
                 "Recipient Input Form",
                 style: ThemeStyle.heading,
               ),
             ),
-            ///calls the textInputForm method to handle the user input
-            _textInputForm("Must enter a First Name", "First Name", Icons.person),
-            _textInputForm("Must enter a Last Name", "Last Name", Icons.person),
-            _textInputForm("Must enter a Adhar Pan Card Number", "Adhar Pan Card Number", Icons.account_balance_wallet),
-            _textInputForm("Must enter a City", "City", Icons.map),
-            _textInputForm("Must enter a State", "state", Icons.map),
-            _textInputForm("Must enter a Phone Number", "Phone Number", Icons.phone),
-          ],
+          ),
+          //first name
+          Padding(
+            padding: const EdgeInsets.only(bottom: 16),
+            child: new TextFormField(
+              decoration: new InputDecoration(
+                hintText: 'First Name',
+                enabledBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.greyColor,
+                    width: 2,
+                  ),
+                  borderRadius: BorderRadius.all(Radius.circular(30)),
+                ),
+                //label text details
+                alignLabelWithHint: true,
+                labelStyle: ThemeStyle.formLabelText,
+                //icon details
+                prefixIcon: Icon(
+                  Icons.person,
+                  size: ThemeStyle.iconSize,
+                  color: ThemeStyle.blackColor,
+                ),
+                //details when selecting that input box
+                focusedBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.primColor,
+                    width: 2,
+                  ),
+                  borderRadius: ThemeStyle.borderRadiusStyle,
+                ),
+                //filling in the color of the inside input box
+                fillColor: ThemeStyle.lightGreyColor,
+                filled: true,
+              ),
+              validator: validateName,
+              onSaved: (String val) {
+                firstName = val;
+              },
+            ),
+          ),
+          // last name
+          Padding(
+            padding: const EdgeInsets.only(bottom: 16),
+            child: new TextFormField(
+              decoration: new InputDecoration(
+                hintText: 'Last Name',
+                enabledBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.greyColor,
+                    width: 2,
+                  ),
+                  borderRadius: BorderRadius.all(Radius.circular(30)),
+                ),
+                //label text details
+                alignLabelWithHint: true,
+                labelStyle: ThemeStyle.formLabelText,
+                //icon details
+                prefixIcon: Icon(
+                  Icons.person,
+                  size: ThemeStyle.iconSize,
+                  color: ThemeStyle.blackColor,
+                ),
+                //details when selecting that input box
+                focusedBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.primColor,
+                    width: 2,
+                  ),
+                  borderRadius: ThemeStyle.borderRadiusStyle,
+                ),
+                //filling in the color of the inside input box
+                fillColor: ThemeStyle.lightGreyColor,
+                filled: true,
+              ),
+              validator: validateName,
+              onSaved: (String val) {
+                lastName = val;
+              },
+            ),
+          ),
+          // card number
+          Padding(
+            padding: const EdgeInsets.only(bottom: 16),
+            child: new TextFormField(
+              decoration: new InputDecoration(
+                hintText: 'Adhar Pan Card Number',
+                enabledBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.greyColor,
+                    width: 2,
+                  ),
+                  borderRadius: BorderRadius.all(Radius.circular(30)),
+                ),
+                //label text details
+                alignLabelWithHint: true,
+                labelStyle: ThemeStyle.formLabelText,
+                //icon details
+                prefixIcon: Icon(
+                  Icons.account_balance_wallet,
+                  size: ThemeStyle.iconSize,
+                  color: ThemeStyle.blackColor,
+                ),
+                //details when selecting that input box
+                focusedBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.primColor,
+                    width: 2,
+                  ),
+                  borderRadius: ThemeStyle.borderRadiusStyle,
+                ),
+                //filling in the color of the inside input box
+                fillColor: ThemeStyle.lightGreyColor,
+                filled: true,
+              ),
+              validator: validateCard,
+              onSaved: (String val) {
+                cardNumber = val;
+              },
+            ),
+          ),
+          // city
+          Padding(
+            padding: const EdgeInsets.only(bottom: 16),
+            child: new TextFormField(
+              decoration: new InputDecoration(
+                hintText: 'City',
+                enabledBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.greyColor,
+                    width: 2,
+                  ),
+                  borderRadius: BorderRadius.all(Radius.circular(30)),
+                ),
+                //label text details
+                alignLabelWithHint: true,
+                labelStyle: ThemeStyle.formLabelText,
+                //icon details
+                prefixIcon: Icon(
+                  Icons.map,
+                  size: ThemeStyle.iconSize,
+                  color: ThemeStyle.blackColor,
+                ),
+                //details when selecting that input box
+                focusedBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.primColor,
+                    width: 2,
+                  ),
+                  borderRadius: ThemeStyle.borderRadiusStyle,
+                ),
+                //filling in the color of the inside input box
+                fillColor: ThemeStyle.lightGreyColor,
+                filled: true,
+              ),
+              validator: validateCity,
+              onSaved: (String val) {
+                city = val;
+              },
+            ),
+          ),
+          // state
+          Padding(
+            padding: const EdgeInsets.only(bottom: 16),
+            child: new TextFormField(
+              decoration: new InputDecoration(
+                hintText: 'State',
+                enabledBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.greyColor,
+                    width: 2,
+                  ),
+                  borderRadius: BorderRadius.all(Radius.circular(30)),
+                ),
+                //label text details
+                alignLabelWithHint: true,
+                labelStyle: ThemeStyle.formLabelText,
+                //icon details
+                prefixIcon: Icon(
+                  Icons.map,
+                  size: ThemeStyle.iconSize,
+                  color: ThemeStyle.blackColor,
+                ),
+                //details when selecting that input box
+                focusedBorder: OutlineInputBorder(
+                  borderSide: BorderSide(
+                    color: ThemeStyle.primColor,
+                    width: 2,
+                  ),
+                  borderRadius: ThemeStyle.borderRadiusStyle,
+                ),
+                //filling in the color of the inside input box
+                fillColor: ThemeStyle.lightGreyColor,
+                filled: true,
+              ),
+              validator: validateState,
+              onSaved: (String val) {
+                state = val;
+              },
+            ),
+          ),
+          // phone number
+          Padding(
+            padding: const EdgeInsets.only(bottom: 16),
+            child: new TextFormField(
+                decoration: new InputDecoration(
+                  hintText: 'Phone Number',
+                  enabledBorder: OutlineInputBorder(
+                    borderSide: BorderSide(
+                      color: ThemeStyle.greyColor,
+                      width: 2,
+                    ),
+                    borderRadius: BorderRadius.all(Radius.circular(30)),
+                  ),
+                  //label text details
+                  alignLabelWithHint: true,
+                  labelStyle: ThemeStyle.formLabelText,
+                  //icon details
+                  prefixIcon: Icon(
+                    Icons.phone,
+                    size: ThemeStyle.iconSize,
+                    color: ThemeStyle.blackColor,
+                  ),
+                  //details when selecting that input box
+                  focusedBorder: OutlineInputBorder(
+                    borderSide: BorderSide(
+                      color: ThemeStyle.primColor,
+                      width: 2,
+                    ),
+                    borderRadius: ThemeStyle.borderRadiusStyle,
+                  ),
+                  //filling in the color of the inside input box
+                  fillColor: ThemeStyle.lightGreyColor,
+                  filled: true,
+                ),
+                keyboardType: TextInputType.phone,
+                validator: validateMobile,
+                onSaved: (String val) {
+                  phoneNumber = val;
+                }),
+          ),
+          new SizedBox(height: 15.0),
+          // button!
+          new RaisedButton(
+            shape: RoundedRectangleBorder(
+              borderRadius: new BorderRadius.circular(30.0),
+            ),
+            onPressed: _sendToServer,
+            child: Text(
+              "SUBMIT",
+              style: ThemeStyle.buttonLabelText,
+            ),
+            color: ThemeStyle.primColor,
+            padding: const EdgeInsets.symmetric(horizontal: 30, vertical: 15),
+          )
+        ],
+      ),
+    );
+  }
+
+  ///checks if name input is empty or invalid characters
+  String validateName(String value) {
+    String patttern = r'(^[a-zA-Z ]*$)';
+    RegExp regExp = new RegExp(patttern);
+    if (value.length == 0) {
+      return "Name is required";
+    } else if (!regExp.hasMatch(value)) {
+      return "Name must be a-z and A-Z";
+    }
+    return null;
+  }
+
+  ///checks if card input is empty or invalid inputs
+  String validateCard(String value) {
+    String patttern = r'(^[0-9]*$)';
+    RegExp regExp = new RegExp(patttern);
+    if (value.length == 0) {
+      return "Card number is required";
+    } else if (!regExp.hasMatch(value)) {
+      return "Card number must be digits";
+    }
+    return null;
+  }
+
+  ///checks if city input is empty
+  String validateCity(String value) {
+    if (value.length == 0) {
+      return "City is required";
+    }
+    return null;
+  }
+
+  ///checks if state input is empty
+  String validateState(String value) {
+    if (value.length == 0) {
+      return "State is required";
+    }
+    return null;
+  }
+
+  ///checks if phone number input is empty or invalid characters
+  String validateMobile(String value) {
+    String patttern = r'(^[0-9]*$)';
+    RegExp regExp = new RegExp(patttern);
+    if (value.length == 0) {
+      return "Phone number is required";
+    } else if (!regExp.hasMatch(value)) {
+      return "Phone number must be digits";
+    }
+    return null;
+  }
+
+  ///saves data, handles navigation
+  _sendToServer() {
+    if (_key.currentState.validate()) {
+      // No any error in validation
+      _key.currentState.save();
+      Navigator.pushNamed(context, '/finishedpage');
+    } else {
+      // validation error
+      setState(() {
+        _validate = true;
+      });
+    }
+  }
+
+
+  @override
+  Widget build(BuildContext context) {
+    // necessary I think, gets ride of debug banner
+    WidgetsApp.debugAllowBannerOverride = false;
+    ScreenUtil.init();
+    ScreenUtil.init(width: 750, height: 1334, allowFontScaling: true);
+    return MaterialApp(
+      home: new Scaffold(
+        appBar: _appBar(),
+        resizeToAvoidBottomPadding: true,
+        body: new SingleChildScrollView(
+          child: new Container(
+            margin: new EdgeInsets.all(15.0),
+            child: new Form(
+              key: _key,
+              autovalidate: _validate,
+              child: _formUI(),
+            ),
+          ),
         ),
       ),
     );
